refactoring(event-store) move event store cache to decorator class
