+++ module/event-sourcing/src/Resources/config/services.yml
@@ -7,7 +7,12 @@ services:
     Ergonode\EventSourcing\Infrastructure\:
         resource: '../../Infrastructure/*'
 
-    Ergonode\EventSourcing\Infrastructure\DomainEventStoreInterface: '@Ergonode\EventSourcing\Infrastructure\Store\DbalDomainEventStore'
+    Ergonode\EventSourcing\Infrastructure\DomainEventStoreInterface:  '@Ergonode\EventSourcing\Infrastructure\Store\DbalDomainEventStore'
+    Ergonode\EventSourcing\Infrastructure\DomainEventStorageInterface:  '@Ergonode\EventSourcing\Infrastructure\Storage\DbalDomainEventStorage'
+    Ergonode\EventSourcing\Infrastructure\Storage\CacheDomainEventStorage:
+      decorates: Ergonode\EventSourcing\Infrastructure\DomainEventStorageInterface
+      arguments:
+        - '@Ergonode\EventSourcing\Infrastructure\Storage\CacheDomainEventStorage.inner'
     Ergonode\EventSourcing\Infrastructure\Provider\DomainEventProviderInterface: '@Ergonode\EventSourcing\Infrastructure\Provider\DomainEventProvider'
     Ergonode\EventSourcing\Infrastructure\Bus\SymfonyCommandBus:
         arguments:
